Input Method Editor (IME) event handling
This update improves the handling of IME events by adding an additional state to track if the input method editor has started. When a Preedit event with a cursor is received, this triggers the start of the IME and when a Commit event is received, this signals the end of the IME. Moreover, the subsequent disabled state no longer triggers a CompositionEnd event. This results in a more robust handling of IME events. Other minor changes include modifications to text insertion and logging behavior.